prev update

+++ src/pages/customers/add-customer.tsx
@@ -188,7 +188,7 @@ export function AddCustomer() {
                     {...registerUpgraded("ServiceCharge", {
                       required: "Service Charge is required",
                       validate: {
-                        positiveNumber: (value) => parseInt(value + "") >= 5,
+                        positiveNumber: (value) => parseFloat(value + "") >= 5,
                       },
                     })}
                   />
